rdpq_debug: fix tile validation by checking only tiles used by the combiner
The previous code was too simplistic; we need to actually look into the
combiner configuration (either 1 cycle or 2 cycles) to check which
tiles are being accessed. It is possible for instance for a 2-step
combiner to only refer to TEX1 in the first cycle, and that means
that the first tile used by the primitive (that'd be TEX0) is not
used and must not be validated.