rdpq_tex_blit: fix various bugs with flipping and source rects
When specifying a source rectangle (subset of the whole surface),
flipping was totally broken. The previous formulas did not account
for this case correctly, in none of the various code paths.

This was true for both flipping via the flip_x/flip_y variables,
and flipping via negative scale.

This patch reworks all the various coordinate formulas to correctly
fix these cases.