n64sym: avoid truncation of line numbers or function offsets
In case either of them is too big to fit 64 KiB (so for functions longer
than 64 KiB of text size, or files with more than 64 KiB source lines),
we just store 0 for the affected symbols, which is better than storing
a wrong (truncated) number. Also not worth making the symbol file larger
to store larger numbers, as such large files are very rare on N64.