rdpq: fix an important heisenbug with blocks that causes flickering
Prevent the same RDP buffer to be re-scheduled twice in sequence,
because most of the times, the buffer can be written by RSP so it could
cause race conditions. See code comment for more information.

This causes visible flickering and even crashes when the same rspq
blocks (containing RDP commmands) was being scheduled many times in
a row, intermixed with other RSP commands.