mixer: avoid forwarding the overread to the waveform callback
Because the RSP mixer ucode does overread (that is, it can read a bit
past the waveform end), we need to make sure that those extra bytes
are 0 (in case of a non looping waveform). Up until now, this was left
to the waveform callback to implement.

We can do better: just like for the looping case, we can have the mixer
handle this internally by clearing up the extra samples, calling the
callback only for actual samples. This better matches the waveform
callback API too, and makes things easier.

Moreover, the overread now is 100% an implementation detail of the mixer.