rdpq_attach: fix Z buffer clearing for non common buffer sizes
The code that computed the RSP DMA hardware register was wrong, in
addition to covering only a few possibilities. For instance, it computed
the wrong value for a 424x240 buffer, and failed for 428x240.

The new code does now correctly calculate the correct RSP DMA size for
both (0x34eff and 0x3bd5f respectively).