n64.mk: activate -ftrivial-auto-var-init
This makes sure all local variables are initialized with some random
pattern. It helps debugging mistakes on usage of uninitialized
variables, and even though it is a debugging function, it does not
affect code size in a significant way.

Also improve the inspector screen to explicitly mention usages of
uninitialized variables when we can detect it (eg: pointers).