n64sys: emit warning in debug log if expansion pak is used without check
Given the fact that expansion pak is not ubiquous, all well-behaved
applications are expected to check the presence of the expansion pak if
needed, to provide a readable error message to the user rather than
crashing for allocation failure.

In a perfect world, we would design libdragon so that usage of expansion
pak is an opt-in, which also generate an error screen. Unfortunately,
this is not possible anymore for backward compatibility, it'd break far
too many things. So we opt for something inbetween:

 * We record if the application called is_memory_expanded() or
   get_memory_size(), assuming that if it did, the developer is doing
   their homework and provide error messages already.
 * We emit a warning message for the developer (in the debug log)
   if more than 4 MiB are allocated, but none of the previous function
   was called.
 * We add a function to provide a simple error screen for people that
   don't want to bother too much: assert_memory_expanded(), which simply
   asserts (still much better experience for the end user than a memory
   crash)