dragonfs: deprecate relative path / current directory support
DragonFS offers a half-baked support for the concept of a current working
directory and relative paths. This is only available via its direct
API, since the global syscalls only work with absolute paths.

The need for a current working directory concept has been ruled out as
not useful, especially for DragonFS. Applications can still perform their
own walking of subdirectories without needing to use a "chdir" syscall.

So we deprecate these functions, as they appear to also be very rarely
used if ever.