system: add real stat() syscall instead of falling back to fstat
Up until now, the stat syscall was implemented by opening the file
and calling fstat on it. This is a bit wasteful as it will also always
call an allocation for the file handle.

Add a real stat() syscall at the filesystem level. This can be useful
mainly to query for a size of a file before opening it.