system: add internal findnext2 filesystem syscall
filesystem's findnext can be hard sometimes to implement because it
does not receive the name of the directory being walked: it can only
use dir->d_cookie (uint32_t) as state, which can be not enough.

Add a new internal findnext2 that also receives a path, and deprecated
findnext, though it'll still work for backward compatibility.