mkfont: first version backported from preview branch
This is a huge tool to convert TTF files to the font64 format, as used
by the rdpq_font library that will be merged soon. This tool is very fast
and supports many details of the TTF/OTF format, plus is able to find
the optimal encoding for the font on N64 among multiple different formats.

Documentation for the tool is available here:
https://github.com/DragonMinded/libdragon/wiki/Mkfont