rdpq: add rdpq_font/paragraph/text library to trunk
This is a vast, fully featured text library finally landing to trunk.
It supports loading and drawing the new font64 format which supports
a wide amount of different fonts rasterized to textures for RDP, with
several different codepaths for optimized rendering.

It includes a full layout engine with advanced features such as
alignment, kerning, word wrapping, and also layout preparation for
faster rendering.

It offers a simple printf-like high-level API (rdpq_text_printf) to
be able to easily draw text on screen.

Moreover, two modern, free builtin fonts are shipped as builtins for
debugging purposes (a monospace one, and a variable space one).