rspq: add workaround for Ares bug causing RSP crashes
The code in rspq_flush_internal avoids a race condition but assumes
perfect CPU/RSP interleaving, which at least Ares cannot guarantee
with the JIT. Adding a branch happens to fix it as per the current
JIT implementation but it's of course fragile.

The correct solution for emulators is to always sync CPU/RSP when
SP_STATUS is accessed, which also makes sense if you think of it.