dma: use 64-bit vaddr in io_read/io_write
This allows to access a wider portion of the PI address space via
the I/O functions, also simplifying about the readability.

Also rewrite a bit the implementations to improve a lot interrupt
latency (don't wait for DMA to finish while keeping interrupts disabled),
and also make it slightly faster.