mksprite: improve greyscale conversion when requesting I4/I8
Before this commit, I4/I8 were really working only on input PNGs which
were already grayscaled. The effect with RGB PNGs was that only the red
channel was kept, which is not what the user wants.

This commits introduces a proper greyscale conversion when the input is
not greyscaled already.

Also, we fix LOD calculation for I4 sprites which was previously not
implemented.