mksprite: autodetect 32-bit PNGs which are actually already quantized
Before this commit, in case of autodetection (no explicit format
specified), mksprite would check if the PNG had an explicit color palette,
and if so, select wither CI4 or CI8 (depending on the number of colors).

It turns out many modern authoring tools (including the widely used
Aseprite) never generate a PNG with a color palette, but instead generate
32-bit PNGs, even though the sprite is actually drawn via a palette. The
actual palette is reconstructed on load.

This commit makes mksprite smarter and detect such PNGs automatically,
so that CI4/CI8 will also be used for them.