mksprite: don't look at alpha channel during quantization
Exoquant tries to quantize taking alpha into account, by premultiplying
it into the colors. This is wrong for N64 because CI4/CI8 don't have
alpha blending anyway, so even if alpha is used in the input picture,
it should be ignored during quantization as it will not be preserved
anyway.