mksprite: allow compressing sprites even while using pipe mode
Currently, when invoking mksprite in pipe mode (stdin/stdout) it is not
possible to compress the sprites. With this commit, it becomes possible
thanks to the use of a temporary file to use before compression.

Also add MKSPRITE_INFN/MKSPRITE_OUTFN for the purpose of giving a name
to stdin/stdout files for logging purposes.