tools: refactor to add a FILE-based compression function for assets
The previous asset_compress_mem is now renamed to asset_compress_mem_raw
and is used to compress an asset in memory. We also add a new
asset_compress_mem that writes the compressed file to disk.

Take also the chance to refactor a bit the code to remove duplications.