dragonfs: add a real stat() function instead of falling back to fstat
Up until now, DragonFS did not have a proper implementation of
the stat syscall, instead falling back to fstat. This is a bit wasteful as
it will also always call an allocation for the file handle.

A change I made that wasn't present prior is that you can now
query directories. This can be useful to query for the type of
directory entry it is (directory, regular file).