Remove nx/ny fields of rdpq_blitparms_t (#791)
Both fields are completely unused outside of rotated blits. nx is also broken for rotated blits.

This is technically a breaking change, but we feel it's important as the fields were merged by mistake to trunk as they are basically broken right now. We believe nobody is using them at the moment.