dragonfs: remove fixed 256-byte size for inodes
This was one of the last relics of dragonfs. inodes (directory_entry_t)
were fixed in size (256 bytes) and also aligned to 256 bytes in ROM.

This refactoring removes this limitation and packs inodes as much as
possible (with a variable size depending on filename). We still
read 256 bytes at runtime, but at least we don't store them.

This should reduce the overhead quite a bit in case of thousands of
files. For reference, gamejam2024 saves 351 KiB