build-toolchain.sh: more portable make invocation
this commit adds a variable that is used to store how to invoke
GNU Make. this is necessary for systems such as the BSDs which
typically use their own make implementation that is not compatible
with the same extensions as GNU Make, and often install the GNU
one as "gmake".

for existing users, nothing changes. BSD users can now invoke with:
MAKE=gmake ./build-toolchain.sh