rspq: remove RSPQCmd_RdpWaitIdle from common code
After years of experience, it would seem this is not very useful to save
imem size in overlays, as waiting for RDP is a pretty specific operation
that is seldom performed. Move it to the rdpq overlay, where it is
instead used.

This also remove quite a bit of common code in case profiling is enabled.